System Chat Improvemts
Blocks all system chat for players with the exception of messages leading with POTATO (e.g. admin messages). Retains full system chat for admins.

Was unable to host a server to properly test this change, however, I was able to properly block and allow the chats by removing the ! in the first if.