DUI settings, now in permanent form (#356)
* DUI settings, now in permanent form

I have no idea if this is the correct format for these....sorry....I don't SQF. I had to do a bunch of find and replace to get it looking correct-ish. And I also can't test this because I'm on a laptop without ARMA installed. Let me know what parts of the formatting are wrong and I can submit fix commits.

* Update BW_Settings.sqf

Co-authored-by: PabstMirror <pabstmirror@gmail.com>