Format Update + Value Tweaks
Standardized code blocks against format on the wiki (https://community.bistudio.com/wiki/Arma_3:_CfgLights) to increase readability and cleanliness.

Also tweaked flashlight mode values in order to nerf the default mode and buff the beam mode. The default mode was as effective as the beam mode (both approx effective to 155m). Now, default mode is effective to 125m and the beam is effective up to 205m.