PTV GL Sight Fix + Structure Cleanup
This patch should fix whatever caused the GL sights to break on the K-M/90 R GL and the G-M/90M GL rifles.

This also brings the patch structure in-line with the current way that PTV restructured their classes (I think).

This should be verified, but I'm like 85% sure this works and fixes the issue at hand.