ACE Wake Up for Non-stable Patients
Adds a flat, non-adjustable wake-up chance for unconscious and non-stable player patients. This is intended to decrease the brutality of going uncon and having to wait for your teammates to check on you.

Wants for this feature:

- Add a %chance to go back unconscious if the patient is not stable and/or has low blood volume. This will highly incentivize players to receive medical attention before running off despite stabilizing themselves.

I've written some pseudo-code so as to show what the above concept might look like.