Cleanup side/systemChat (#273)
* Cleanup side/systemChat

Reduce messages shown in chat (less distractions, better stream/recordings)
- Prevents accident use of side/global text chat
- Players can still send messages with force prefix (e.g. `force blu ready`)
- Reduces some "Player connecting" spam (will be improved in 2.02)

* System Chat Improvemts

Blocks all system chat for players with the exception of messages leading with POTATO (e.g. admin messages). Retains full system chat for admins.

Was unable to host a server to properly test this change, however, I was able to properly block and allow the chats by removing the ! in the first if.

* Chat warning fix

Reformats the chat warning to prevent it from being blocked.

* Only block player connecting systemChat, add setting

* Update XEH_postInit.sqf

* Update XEH_postInit.sqf

* Add getVar to disable in case this causes problems

---------

Co-authored-by: Alablm <aidan197@hotmail.com>